LatentWorker: locks and notifiers should not be recreated on reconfig
In case of the notifier, if it is recreated on reconfig while a build is
running, the master would wait for the old notifier to be notifying, but
it never would because it has been replaced and the notification of the
build's end would only take place on the new notifier.

This particular behavior would prevent a master from stopping in the
following sequence of events:
- start master with latent docker worker configuration that prevents the
  docker worker from connecting to the master (e.g. wrong master FQDN),
  thus builds on this worker get stuck in the "preparing worker" state
- start a build, which gets stuck in the "preparing worker" state
- reconfigure master with correct master FQDN
- restart master

The master would now not stop but sit there waiting for the build to
finish. The build has actually finished (it has been cancelled due to
the restart), but the notification is not sent correctly to the master's
shutdown routine.

Upstreamed at https://github.com/buildbot/buildbot/pull/5599