gerrit: Use the actual git branch for the change branch attribute
Gerrit change sources used a different definition of what the branch
attribute of a change source means. Rather than using the actual git
branch they invented a string that is common for all events coming from
a Gerrit change so that these changes could be grouped in the UI.

Unfortunately this breaks event processing downstream such as event
filtering.